Removed warnings in build logs. (changed a few data types is all)

+++ Game.h
@@ -59,6 +59,6 @@ private:
 
 	//create camera
 	std::vector<std::shared_ptr<Camera>> cameras;
-	UINT currentCameraIndex;
+	size_t currentCameraIndex;
 };
